+++ apps/spaces/src/components/QuizViewLayout/components/QuestionTableDraggableRow.tsx
@@ -1,0 +1,183 @@
+import { FunctionComponent, memo } from "react";
+import { flexRender, Row } from "@tanstack/react-table";
+import { useSortable } from "@dnd-kit/sortable";
+import { CSS } from "@dnd-kit/utilities";
+import { ActionTooltip, defaultTheme, EditIcon, styled } from "@horizontal-org/shira-ui";
+import { QuizQuestion } from "../../../store/slices/quiz";
+import { QuestionTableActionsMenu } from "./QuestionTableActionsMenu";
+
+interface Props {
+  row: Row<QuizQuestion>;
+  duplicatingQuestionId: string | null;
+  onEditQuestion: (questionId: string) => void;
+  onDuplicateQuestion: (questionId: string) => void;
+  onDeleteQuestion: (questionId: string) => void;
+  editTooltip: string;
+  duplicateTooltip: string;
+  deleteTooltip: string;
+}
+
+const QuestionTableDraggableRowComponent: FunctionComponent<Props> = ({
+  row,
+  duplicatingQuestionId,
+  onEditQuestion,
+  onDuplicateQuestion,
+  onDeleteQuestion,
+  editTooltip,
+  duplicateTooltip,
+  deleteTooltip,
+}) => {
+  const {
+    attributes,
+    listeners,
+    setNodeRef,
+    transform,
+    transition,
+    isDragging,
+  } = useSortable({
+    id: row.original.question.id,
+    animateLayoutChanges: () => false,
+  });
+
+  const isDuplicatingThisQuestion = duplicatingQuestionId === row.original.question.id;
+
+  return (
+    <Tr
+      ref={setNodeRef}
+      style={{
+        transform: CSS.Transform.toString(transform),
+        transition,
+      }}
+      $dragging={isDragging}
+      id={`question-item-${row.original.question.id}`}
+    >
+      {row.getVisibleCells().map((cell) => {
+        if (cell.column.id === "drag") {
+          return (
+            <Td key={cell.id}>
+              <HandleButton
+                type="button"
+                {...attributes}
+                {...listeners}
+                disabled={isDuplicatingThisQuestion}
+              >
+                {flexRender(cell.column.columnDef.cell, cell.getContext())}
+              </HandleButton>
+            </Td>
+          );
+        }
+
+        if (cell.column.id === "actions") {
+          const questionId = row.original.question.id;
+
+          return (
+            <Td key={cell.id}>
+              <ActionsCell>
+                <ActionTooltip content={editTooltip}>
+                  <ActionButton
+                    id={`edit-button-${questionId}`}
+                    type="button"
+                    aria-label={editTooltip}
+                    onClick={() => onEditQuestion(questionId)}
+                  >
+                    <EditIcon />
+                  </ActionButton>
+                </ActionTooltip>
+
+                <QuestionTableActionsMenu
+                  questionId={questionId}
+                  duplicateLabel={duplicateTooltip}
+                  deleteLabel={deleteTooltip}
+                  disabled={isDuplicatingThisQuestion}
+                  onDuplicate={() => onDuplicateQuestion(questionId)}
+                  onDelete={() => onDeleteQuestion(questionId)}
+                />
+              </ActionsCell>
+            </Td>
+          );
+        }
+
+        return (
+          <Td key={cell.id}>
+            {flexRender(cell.column.columnDef.cell, cell.getContext())}
+          </Td>
+        );
+      })}
+    </Tr>
+  );
+};
+
+export const QuestionTableDraggableRow = memo(QuestionTableDraggableRowComponent);
+
+const Td = styled.td`
+  background: ${defaultTheme.colors.light.white};
+  padding: 14px 16px;
+  vertical-align: middle;
+  box-sizing: border-box;
+  width: inherit;
+  font-size: inherit;
+  border-bottom: 1px solid ${defaultTheme.colors.light.paleGrey};
+`;
+
+const Tr = styled.tr<{ $dragging?: boolean }>`
+  position: relative;
+
+  &:hover td {
+    background: ${defaultTheme.colors.light.paleGrey};
+  }
+
+  &:last-child td {
+    border-bottom: none;
+  }
+
+  ${(props) => props.$dragging && `
+    z-index: 2;
+
+    & td {
+      background: ${defaultTheme.colors.light.white};
+    }
+  `}
+`;
+
+const HandleButton = styled.button`
+  all: unset;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  cursor: grab;
+  touch-action: none;
+
+  &:active {
+    cursor: grabbing;
+  }
+
+  &:disabled {
+    cursor: not-allowed;
+  }
+`;
+
+const ActionsCell = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 8px;
+`;
+
+const ActionButton = styled.button`
+  background: none;
+  border: none;
+  padding: 8px;
+  cursor: pointer;
+  border-radius: 4px;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+
+  &:hover:not(:disabled) {
+    background: ${defaultTheme.colors.light.paleGreen};
+  }
+
+  &:disabled {
+    cursor: not-allowed;
+    opacity: 0.7;
+  }
+`;
