+++ packages/shira-ui/src/components/Table/index.ts
@@ -1,5 +1,4 @@
 export * from './Table'
-export * from './SortableTable'
 export * from './components/TableActions'
 export * from './components/TableCheckbox'
 export * from './components/Pagination'
